requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
bootstrap4==0.1.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-annoying==0.10.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-leaflet==0.27.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
websocket-client==0.57.0
wrapt==1.12.1
zope.interface==5.1.0
